Commit for summary results of HGT filtering on model BB long run w/ 9900 trees
No code committed, YET! Included are text files each representing a column in a matrix, of which trees for this particular model passed the given HGT constraint. Also contains a text file summary with number of trees passing each of these 16 constraints. Although this summary file isn't the best metric for rank/order and analysis of these HGTs like the Jaccard Index we plan to use, it's still useful nonetheless. Plots to come later! Thanks, Jack