Adding HGT constraint data frame/matrix for model BB.
Adding a data frame/matrix for PhyloBayes model BB long with 9900 trees, indexed as unqiue and ordered rows; with 16 HGTs as columns and a binary numeric value of 0 or 1 indicating a tree passed a given constraint.

TBD: This is to be used for clustering/similarity/jaccard analysis to investigate the similarity between HGT constraints as applied to this model and to investigate their restrictive or permissive power.